This project is a simple communication and information storage application. The interface was created with the tkinter library.
This project is a simple communication and information storage application. It has announcements, projects, lessons and chat section. You can sort your projects by dates and see how much time is left, and there is an extra section to add special notes to each project. In the course section, you can add lesson hours on any day of the week and at any time you want.
If the server is open in the chat section, you can communicate instantly via text with people active on the server.
There is a special button for Background. There is a list of images in the code. The user can add the background images he wants by copying the file path of his own png images and adding them to this list.